# BUG: Fix binary_repr for negative numbers