--- a/numpy/core/tests/test_deprecations.py
+++ b/numpy/core/tests/test_deprecations.py
@@ -579,19 +579,45 @@ class TestArrayDataAttributeAssignmentDeprecation(_DeprecationTestCase):
 
 
 class TestLinspaceInvalidNumParameter(_DeprecationTestCase):
-    """Argument to the num parameter in linspace that cannot be 
+    """Argument to the num parameter in linspace that cannot be
     safely interpreted as an integer is deprecated in 1.12.0.
 
-    Argument to the num parameter in linspace that cannot be 
-    safely interpreted as an integer should not be allowed. 
+    Argument to the num parameter in linspace that cannot be
+    safely interpreted as an integer should not be allowed.
     In the interest of not breaking code that passes
     an argument that could still be interpreted as an integer, a
-    DeprecationWarning will be issued for the time being to give 
+    DeprecationWarning will be issued for the time being to give
     developers time to refactor relevant code.
     """
     def test_float_arg(self):
         # 2016-02-25, PR#7328
         self.assert_deprecated(np.linspace, args=(0, 10, 2.5))
+
+
+class TestBinaryReprInsufficientWidthParameterForRepresentation(_DeprecationTestCase):
+    """
+    If a 'width' parameter is passed into ``binary_repr`` that is insufficient to
+    represent the number in base 2 (positive) or 2's complement (negative) form,
+    the function used to silently ignore the parameter and return a representation
+    using the minimal number of bits needed for the form in question. Such behavior
+    is now considered unsafe from a user perspective and will raise an error in the future.
+    """
+
+    def test_insufficient_width_positive(self):
+        args = (10,)
+        kwargs = {'width': 2}
+
+        self.message = ("Insufficient bit width provided. This behavior "
+                        "will raise an error in the future.")
+        self.assert_deprecated(np.binary_repr, args=args, kwargs=kwargs)
+
+    def test_insufficient_width_negative(self):
+        args = (-5,)
+        kwargs = {'width': 2}
+
+        self.message = ("Insufficient bit width provided. This behavior "
+                        "will raise an error in the future.")
+        self.assert_deprecated(np.binary_repr, args=args, kwargs=kwargs)
 
 
 class TestTestDeprecated(object):