Addresses issue in #7168 in which passing in negative numbers with insufficient `width` parameters resulted in completely incorrect two's complement representations (e.g. all zero's) or even an error. Now, it will return the two's complement with width equal to the minimum number of bits needed to represent the complement.