@gfyoung 2016-02-03
As noted in my commit message, using Python built-in `bin` provides a nice speed up to what is already a pretty high-performing function:

```python -m timeit -n 10000 -s "import numpy as np" "np.binary_repr(123456789987654321)"```

On master: 4.07 usec per loop
On PR      : 2.99 usec per loop

```python -m timeit -n 10000 -s "import numpy as np" "np.binary_repr(-123456789987654321)"```

On master: 4.19 usec per loop
On PR      : 1.85 usec per loop

```python -m timeit -n 10000 -s "import numpy as np" "np.binary_repr(10, width=60)"```

On master: 2.33 usec per loop
On PR      : 2.12 usec per loop

```python -m timeit -n 10000 -s "import numpy as np" "np.binary_repr(-123456789, width=5)"```

On master: `KeyError`
On PR      : 3.09 usec per loop