@gfyoung 2016-02-17
@charris : if there are no issues, can this be merged now since #7194 was merged in?  Or can someone else take a look at this?