@ahaldane 2016-02-17
If it is not possible to represent the number given the supplied width, would it be better to just raise an exception? Silently returning a value with a width different from what was requested seems dangerous.