@gfyoung 2016-02-17
It's for backwards compatibility reasons.  That's what is done with positive numbers, so the behavior should be consistent with negative numbers.