@ahaldane 2016-02-17
I see. 

But I still feel it's dangerous, and the positive case could be considered a bug. There are no unit tests for it. Maybe we should take this opportunity to break backward compatibility, and raise an exception in that case? `base_repr` doesn't have this problem, by the way.

Can anyone more experienced on backward compatibility problems comment?