@gfyoung 2016-02-17
Well, I would beg to differ in terms of the testing.  If you look at the test suite, I explicitly wrote one for insufficient widths to make it go the behavior I coded it to do.  However, I can understand why you might be "surprised" if you suddenly got a representation with a width larger than you had requested.  TBH I don't quite understand why the function was written as it was previously, so another person's input would be great as well!