@ahaldane 2016-02-17
Oh, I meant there were no *existing* tests for extending the width past the user's request, meaning it may be unintentional. I do see you added some though :)