@gfyoung 2016-02-19
Does anyone have anything to add to this?  At this point, I'm still leaning towards maintaining backwards compatibility, but other opinions are welcome!