@seberg 2016-03-05
I have to admit I have a slight preference for just erroring out if the width is not enough to represent the twos complement for negative numbers. Unless that is a backwards incompatible break?