@gfyoung 2016-03-06
@seberg : There would be a backwards compatibility break because that behavior is used for positive numbers, so for consistency, it should be used for negative numbers as well.