@seberg 2016-03-06
Hmmmm, good point. On the other hand, the user suddenly has to know whether his number was positive or negative to make sense of the result. Unrelated, I think we could mention `bin` in the doc string. Since the usefulness of this function was probably partially when `bin` did not exist.