@gfyoung 2016-03-06
1) Given that the binary representation (without knowing anything else) could have been for a positive or negative number, what do you suggest for the first point then?

2) I did not realize there was a time when `bin` did not exist.  I agree that that should be added to the documentation.