@seberg 2016-03-06
I don't know, but I assumed since it was not used in the function ;). But we are going back to python 1.x for this type of functions I think. Not sure, I guess we would have to deprecate it to throw an error also for positive numbers and maybe just throw an error for negative, since it mostly didn't work anyway.