@seberg 2016-03-06
Uhh, brand new: "New in version 2.6." :)