@gfyoung 2016-03-06
You mean deprecate behavior when the `width` parameter is insufficient?