@seberg 2016-03-06
Yeah, just because if we would prefer error for negative + insufficient width it seems more right to also do it for positive (and I don't really see a good reason to extend width silently). Still thinking about it, but I think I would like it best. Though this is a bug fix foremost I guess.