@gfyoung 2016-03-06
@seberg : Added the deprecation, and the tests are passing.  Can you merge this?