@seberg 2016-03-06
Anyway, thanks. Will leave it sitting a bit in case someone disagrees.