@gfyoung 2016-03-06
Sure thing.  I'll ping again in about a week to see if anyone else has anything to add.  Would be nice to land this at some point.