@ahaldane 2016-03-06
I, for one, am satisfied. Merge at will!