@gfyoung 2016-03-13
@seberg , @ahaldane : it has been a week, and I don't see any complaints.  Can this be merged?