@njsmith 2016-03-13
I haven't reviewed in depth, but it looks like @ahaldane and @seberg both have and are happy and no-one else has spoken up to object, so merging.