@gfyoung 2016-03-13
Thanks @njsmith !  Also thanks to @ahaldane and @seberg for going over this thoroughly.