@njsmith 2016-03-13
@gfyoung: the bottleneck is generlaly someone having the time to read through and think about things; once the decision is made then clicking the button isn't a big deal. The practical consequence is that if the merge conflicts are just a trivial issue with the release notes, then please don't waste your time rebasing over and over... just explain the merge conflict being trivial in the ping so it's clear that the relevant content is still reviewable, and then once you get the :+1: you can rebase once then :-)